StopIndexing work 1

+++ src/main/java/searchengine/services/IndexingServiceImpl.java
@@ -34,44 +34,42 @@ public class IndexingServiceImpl implements IndexingService
     ExecutorService executor;   //29
 
     @Override
-    public void startIndexing()
+    public boolean startIndexing()
     {
         // Блок Б1
         System.out.println("\nЗапущен метод startIndexing");
-        isIndexingStarted = true;
 
-        // Блок Б2 с удалением сайта из таблицы sites
-        for (Site site:sites.getSites())
+        if(!isIndexingStarted)
         {
-            System.out.println(site);
-            Iterable<searchengine.model.Site> siteIterable = siteRepository.findAll();
-            for (searchengine.model.Site siteDB : siteIterable)
-                {
-                    if(site.getUrl().equals(siteDB.getUrl()))
-                        {
+            isIndexingStarted = true;
+
+            // Блок Б2 с удалением сайта из таблицы sites
+            for (Site site : sites.getSites()) {
+                System.out.println(site);
+                Iterable<searchengine.model.Site> siteIterable = siteRepository.findAll();
+                for (searchengine.model.Site siteDB : siteIterable) {
+                    if (site.getUrl().equals(siteDB.getUrl())) {
                         siteRepository.delete(siteDB);
                         System.out.println("\nВыполнено удаление сайта: " + site.getUrl());
-                        }
+                    }
                 }
-        }
+            }
 
-        // Блок Б3.2/2
-        for (Site site:sites.getSites())
-        {
-            searchengine.model.Site siteDB = new searchengine.model.Site();
-            siteDB.setName(site.getName());
-            siteDB.setUrl(site.getUrl());
-            siteDB.setStatus(StatusType.INDEXING);
-            siteDB.setStatusTime(new Date());
-            siteRepository.save(siteDB);
+            // Блок Б3.2/2
+            for (Site site : sites.getSites()) {
+                searchengine.model.Site siteDB = new searchengine.model.Site();
+                siteDB.setName(site.getName());
+                siteDB.setUrl(site.getUrl());
+                siteDB.setStatus(StatusType.INDEXING);
+                siteDB.setStatusTime(new Date());
+                siteRepository.save(siteDB);
 
-            System.out.println("Выполнено сохранение сайта " + site);   //*
-        }
+                System.out.println("Выполнено сохранение сайта " + site);   //*
+            }
 
 //        Stack<RunnableFuture<Boolean>> taskList = new Stack<>();  //29
-        Iterable<searchengine.model.Site> siteIterable = siteRepository.findAll();
-        for (searchengine.model.Site siteDB : siteIterable)
-        {
+            Iterable<searchengine.model.Site> siteIterable = siteRepository.findAll();
+            for (searchengine.model.Site siteDB : siteIterable) {
 
 //            try {
 //                ForkJoinTask<?> result = new ForkJoinTask<?>().adapt((Runnable) new PageWriter(siteDB));
@@ -86,13 +84,13 @@ public class IndexingServiceImpl implements IndexingService
             System.out.println("\nВыполнено добавление сайта " + siteDB + " в FJP/Future");
             */
 
-            //            /*
-            //Вар.3
-            StartIndexing value = new StartIndexing(siteDB);
-            RunnableFuture<Boolean> futureValue = new FutureTask<>(value);
-            listStartIndexing.add(value);
-            taskList.add(futureValue);
-            System.out.println("\nВыполнено добавление сайта " + siteDB + " в FJP/Future");
+                // /*
+                //Вар.3
+                StartIndexing value = new StartIndexing(siteDB);
+                RunnableFuture<Boolean> futureValue = new FutureTask<>(value);
+                listStartIndexing.add(value);
+                taskList.add(futureValue);
+                System.out.println("\nВыполнено добавление сайта " + siteDB + " в FJP/Future");
 //            */
 
             /*
@@ -107,23 +105,29 @@ public class IndexingServiceImpl implements IndexingService
             taskList.add(futureValue);
             System.out.println("\nВыполнено добавление сайта " + siteDB + " в FJP/Future");
             */
-        }
+            }
 
 //        ExecutorService executor = Executors.newFixedThreadPool(4); // //29
-        executor = Executors.newFixedThreadPool(4); // TODO : Внести количество потоков с привязкой к количеству сайтов
-        taskList.forEach(executor::execute);
+            executor = Executors.newFixedThreadPool(4); // TODO : Внести количество потоков с привязкой к количеству сайтов
+            taskList.forEach(executor::execute);
 
 //        ResultCheckerParse resultCheckerExample = new ResultCheckerParse(taskList);   //1 рабочее
-        resultCheckerExample = new ResultCheckerParse(taskList);  //1 рабочее
+            resultCheckerExample = new ResultCheckerParse(taskList);  //1 рабочее
 //        resultCheckerExample = new ResultCheckerParse();
 
-        executor.execute(resultCheckerExample);
+            executor.execute(resultCheckerExample);
 
-        executor.shutdown();
+            executor.shutdown();
 
 //        isIndexingStarted = false; // Завершать нужно по завершению потоков - RCP
 
-        System.out.println("\nЗавершение метода startIndexing в классе IndexingServiceImpl");
+            System.out.println("\nЗавершение выполнившегося метода startIndexing в классе IndexingServiceImpl");    //*
+            return true;
+        } else
+            {
+                System.out.println("\nISImpl - Ошибка: Индексация уже запущена");   //*
+                return false;
+            }
     }
 
     @Override
